Added a CoordinateHash and CoordinateEqual with overloaded () opeartors so Coordinates can be used in unorederd_maps. Need to redo the PriorityQueue wrapper

+++ Dwarf/Engine.cpp
@@ -63,7 +63,7 @@ void Engine::run()
 		// This will have issues when the game becomes simulation bound!@!@!@!@!@!
 		while (frameTime > 0.0)
 		{
-			float deltaTime = std::min(frameTime, dt);
+			double deltaTime = std::min(frameTime, dt);
 			update(deltaTime);
 
 			frameTime -= deltaTime;
